requirements.txt
audioread==3.0.1
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
decorator==5.1.1
Flask==3.1.0
Flask-SocketIO==5.5.1
gTTS==2.5.4
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.44.0
lxml==5.3.0
MarkupSafe==3.0.2
msgpack==1.1.0
numba==0.61.0
numpy==2.1.3
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
pycparser==2.22
pydub==0.25.1
python-engineio==4.11.2
python-pptx==1.0.2
python-socketio==5.12.1
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
simple-websocket==1.1.0
soundfile==0.13.1
soxr==0.5.0.post1
threadpoolctl==3.5.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wsproto==1.2.0
XlsxWriter==3.2.2
